Fix some issues including typos and the type of the product ID

+++ libsource/src/main/java/com/virtusize/libsource/Virtusize.kt
@@ -348,7 +348,7 @@ class Virtusize(
     }
 
     fun getStoreProductInfo(
-        productId: Long,
+        productId: Int,
         onSuccess: ((StoreProduct?) -> Unit)? = null,
         onError: ((VirtusizeError) -> Unit)? = null
     ) {
